dbt-weather-apoena/pyproject.toml
[project]
name = "dbt-weather-apoena"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dbt-core>=1.10.11",
    "dbt-postgres>=1.9.1",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "ruff>=0.13.0",
]
